@@ -1,14 +1,253 @@
-//
-// Created by igor on 6/16/24.
-//
-#include <sdlpp/video/blend_mode.hh>
-#include "utils/switch_ostream.hh"
+/**
+ * @file blend_mode.cc
+ * @brief Stream operator implementations for enums
+ * @note This file is auto-generated by generate_enum_operators.py
+ */
 
-namespace neutrino::sdl {
-	BEGIN_IMPL_OUTPUT(blend_mode)
-			case blend_mode::NONE: return "NONE";
-			case blend_mode::BLEND: return "BLEND";
-			case blend_mode::ADD: return "ADD";
-			case blend_mode::MOD: return "MOD";
-	END_IMPL_OUTPUT(blend_mode)
-}
+#include <sdlpp/video/blend_mode.hh>
+#include <iostream>
+#include <string>
+#include <algorithm>
+#include <cctype>
+#include <vector>
+
+namespace sdlpp {
+    std::ostream& operator<<(std::ostream& os, blend_mode value) {
+        switch (value) {
+            case blend_mode::none:
+                return os << "none";
+            case blend_mode::blend:
+                return os << "blend";
+            case blend_mode::blend_premultiplied:
+                return os << "blend_premultiplied";
+            case blend_mode::add:
+                return os << "add";
+            case blend_mode::add_premultiplied:
+                return os << "add_premultiplied";
+            case blend_mode::mod:
+                return os << "mod";
+            case blend_mode::mul:
+                return os << "mul";
+            case blend_mode::invalid:
+                return os << "invalid";
+            default:
+                return os << "Unknown_blend_mode(" << static_cast <int>(value) << ")";
+        }
+    }
+
+    std::istream& operator>>(std::istream& is, blend_mode& value) {
+        std::string str;
+        is >> str;
+
+        // Check if hex input is expected based on stream flags
+        if (is.flags() & std::ios::hex) {
+            try {
+                unsigned int int_value = static_cast<unsigned int>(std::stoul(str, nullptr, 16));
+                value = static_cast <blend_mode>(int_value);
+                return is;
+            } catch (...) {
+                is.setstate(std::ios::failbit);
+                return is;
+            }
+        }
+
+        // Convert to lowercase for case-insensitive comparison
+        std::transform(str.begin(), str.end(), str.begin(), ::tolower);
+
+        if (str == "none") {
+            value = blend_mode::none;
+        } else if (str == "blend") {
+            value = blend_mode::blend;
+        } else if (str == "blend_premultiplied") {
+            value = blend_mode::blend_premultiplied;
+        } else if (str == "add") {
+            value = blend_mode::add;
+        } else if (str == "add_premultiplied") {
+            value = blend_mode::add_premultiplied;
+        } else if (str == "mod") {
+            value = blend_mode::mod;
+        } else if (str == "mul") {
+            value = blend_mode::mul;
+        } else if (str == "invalid") {
+            value = blend_mode::invalid;
+        } else {
+            // Try to parse as integer
+            try {
+                int int_value = std::stoi(str);
+                value = static_cast <blend_mode>(int_value);
+            } catch (...) {
+                is.setstate(std::ios::failbit);
+            }
+        }
+
+        return is;
+    }
+
+    std::ostream& operator<<(std::ostream& os, scale_mode value) {
+        switch (value) {
+            case scale_mode::nearest:
+                return os << "nearest";
+            case scale_mode::linear:
+                return os << "linear";
+            default:
+                return os << "Unknown_scale_mode(" << static_cast <int>(value) << ")";
+        }
+    }
+
+    std::istream& operator>>(std::istream& is, scale_mode& value) {
+        std::string str;
+        is >> str;
+
+        // Check if hex input is expected based on stream flags
+        if (is.flags() & std::ios::hex) {
+            try {
+                unsigned int int_value = static_cast<unsigned int>(std::stoul(str, nullptr, 16));
+                value = static_cast <scale_mode>(int_value);
+                return is;
+            } catch (...) {
+                is.setstate(std::ios::failbit);
+                return is;
+            }
+        }
+
+        // Convert to lowercase for case-insensitive comparison
+        std::transform(str.begin(), str.end(), str.begin(), ::tolower);
+
+        if (str == "nearest") {
+            value = scale_mode::nearest;
+        } else if (str == "linear") {
+            value = scale_mode::linear;
+        } else {
+            // Try to parse as integer
+            try {
+                int int_value = std::stoi(str);
+                value = static_cast <scale_mode>(int_value);
+            } catch (...) {
+                is.setstate(std::ios::failbit);
+            }
+        }
+
+        return is;
+    }
+
+    std::ostream& operator<<(std::ostream& os, flip_mode value) {
+        // Check if hex output is requested
+        if (os.flags() & std::ios::hex) {
+            // Output as hex value
+            os << static_cast <unsigned int>(value);
+            return os;
+        }
+
+        // Output as human-readable flag names
+        std::vector <std::string> active_flags;
+
+        if ((static_cast <unsigned int>(value) & static_cast <unsigned int>(flip_mode::none)) != 0) {
+            active_flags.emplace_back("none");
+        }
+        if ((static_cast <unsigned int>(value) & static_cast <unsigned int>(flip_mode::horizontal)) != 0) {
+            active_flags.emplace_back("horizontal");
+        }
+        if ((static_cast <unsigned int>(value) & static_cast <unsigned int>(flip_mode::vertical)) != 0) {
+            active_flags.emplace_back("vertical");
+        }
+
+        // Sort flags lexicographically
+        std::sort(active_flags.begin(), active_flags.end());
+
+        if (active_flags.empty()) {
+            os << "none";
+        } else {
+            bool first = true;
+            for (const auto& flag : active_flags) {
+                if (!first) os << " | ";
+                os << flag;
+                first = false;
+            }
+        }
+        return os;
+    }
+
+    std::istream& operator>>(std::istream& is, flip_mode& value) {
+        std::string str;
+        is >> str;
+
+        value = static_cast <flip_mode>(0);
+
+        // Check if hex input is expected based on stream flags
+        if (is.flags() & std::ios::hex) {
+            try {
+                unsigned int int_value = static_cast<unsigned int>(std::stoul(str, nullptr, 16));
+                value = static_cast <flip_mode>(int_value);
+                return is;
+            } catch (...) {
+                is.setstate(std::ios::failbit);
+                return is;
+            }
+        }
+
+        // Check for explicit hex format (0x...)
+        if (str.substr(0, 2) == "0x" || str.substr(0, 2) == "0X") {
+            try {
+                unsigned int int_value = static_cast<unsigned int>(std::stoul(str, nullptr, 16));
+                value = static_cast <flip_mode>(int_value);
+                return is;
+            } catch (...) {
+                is.setstate(std::ios::failbit);
+                return is;
+            }
+        }
+
+        // Parse as flag names separated by |
+        size_t start = 0;
+        size_t pos = 0;
+
+        while (start < str.length()) {
+            pos = str.find('|', start);
+            std::string flag_name = (pos == std::string::npos) ? str.substr(start) : str.substr(start, pos - start);
+
+            // Trim whitespace
+            size_t first = flag_name.find_first_not_of(' ');
+            size_t last = flag_name.find_last_not_of(' ');
+            if (first != std::string::npos && last != std::string::npos) {
+                flag_name = flag_name.substr(first, last - first + 1);
+            }
+
+            // Convert to lowercase
+            std::transform(flag_name.begin(), flag_name.end(), flag_name.begin(), ::tolower);
+
+            // Match flag name
+            bool found = false;
+            if (flag_name == "none") {
+                value = static_cast <flip_mode>(static_cast <unsigned int>(value) | static_cast <unsigned int>(
+                                                    flip_mode::none));
+                found = true;
+            }
+            if (flag_name == "horizontal") {
+                value = static_cast <flip_mode>(static_cast <unsigned int>(value) | static_cast <unsigned int>(
+                                                    flip_mode::horizontal));
+                found = true;
+            }
+            if (flag_name == "vertical") {
+                value = static_cast <flip_mode>(static_cast <unsigned int>(value) | static_cast <unsigned int>(
+                                                    flip_mode::vertical));
+                found = true;
+            }
+
+            if (!found) {
+                // Try to parse as integer
+                try {
+                    unsigned int int_value = static_cast<unsigned int>(std::stoul(flag_name));
+                    value = static_cast <flip_mode>(static_cast <unsigned int>(value) | int_value);
+                } catch (...) {
+                    is.setstate(std::ios::failbit);
+                    return is;
+                }
+            }
+
+            if (pos == std::string::npos) break;
+            start = pos + 1;
+        }
+
+        return is;
+    }
+} // namespace
